Make all lib functions internal, add contracts for testing
