+++ packages/contracts/src/contracts/current/protocol/Exchange/LibOrder.sol
@@ -55,8 +55,9 @@ contract LibOrder {
     /// @dev Calculates Keccak-256 hash of the order.
     /// @param order The order structure.
     /// @return Keccak-256 EIP712 hash of the order.
-    function getOrderHash(Order order)
-        public view
+    function getOrderHash(Order memory order)
+        internal
+        view
         returns (bytes32 orderHash)
     {
         // TODO: EIP712 is not finalized yet
