+++ packages/contracts/src/contracts/current/protocol/Exchange/MixinAssetProxyDispatcher.sol
@@ -78,7 +78,8 @@ contract MixinAssetProxyDispatcher is
     /// @param assetProxyId Id of the asset proxy.
     /// @return The asset proxy registered to assetProxyId. Returns 0x0 if no proxy is registered.
     function getAssetProxy(uint8 assetProxyId)
-        external view
+        external
+        view
         returns (address)
     {
         IAssetProxy assetProxy = assetProxies[assetProxyId];
