+++ packages/contracts/src/contracts/current/protocol/Exchange/MixinSignatureValidator.sol
@@ -24,6 +24,7 @@ import "./ISigner.sol";
 
 /// @dev Provides MSignatureValidator
 contract MixinSignatureValidator is MSignatureValidator {
+
     enum SignatureType {
         Illegal, // Default value
         Invalid,
@@ -47,7 +48,8 @@ contract MixinSignatureValidator is MSignatureValidator {
         bytes32 hash,
         address signer,
         bytes memory signature)
-        public view
+        internal
+        view
         returns (bool isValid)
     {
         // TODO: Domain separation: make hash depend on role. (Taker sig should not be valid as maker sig, etc.)
@@ -164,8 +166,8 @@ contract MixinSignatureValidator is MSignatureValidator {
     function preSign(
         bytes32 hash,
         address signer,
-        bytes memory signature)
-        public
+        bytes signature)
+        external
     {
         require(isValidSignature(hash, signer, signature));
         preSigned[hash][signer] = true;
