+++ packages/contracts/src/contracts/current/protocol/Exchange/mixins/MSignatureValidator.sol
@@ -30,6 +30,7 @@ contract MSignatureValidator {
         bytes32 hash,
         address signer,
         bytes memory signature)
-        public view
+        internal
+        view
         returns (bool isValid);
 }
